+++ srcs/client.cpp
@@ -1,4 +1,4 @@
-#include "client.hpp"
+#include "webserv.hpp"
 
 Client::Client()
 {
