+++ srcs/response_redirection.cpp
@@ -1,10 +1,8 @@
-#include "response.hpp"
-#include "client.hpp"
+#include "webserv.hpp"
 
 void		Response::makeRedirectionResponse()
 {
-	int ret;
-	initResponse();
+	this->raw_response.clear();
     
     addFirstLine(this->location->getRedirectReturn());
     addDate();
@@ -12,5 +10,6 @@ void		Response::makeRedirectionResponse()
     addLocation();
 	
     this->client->setStatus(RESPONSE_MAKE_DONE);
+    this->is_disconnect_immediately = true;
     return ;
 }
