+++ tests/sample/test.conf
@@ -5,19 +5,55 @@
 server
 {
 	server_name default_server
-	listen 8280 127.0.0.1			# 7번라인 8번라인 반드시 연달아와야합니다.
+	listen 8183 127.0.0.1			# 7번라인 8번라인 반드시 연달아와야합니다.
 
-	location /					# / 로케이션은 반드시 존재해야합니다. (디폴트서버)
+	location /						# / 로케이션은 반드시 존재해야합니다. (디폴트서버)
 	{
-		allow_methods GET
-		root ./tester/YoupiBanane/
-		index youpi.bad_extension
+		error_page 404 ./tests/sample/error404.html
+		error_page 405 ./tests/sample/error405.html
+		allow_methods GET POST
+		request_max_body_size	2048
+		root ./tests/sample/
+		index index.html index2.html
+		auto_index on
+		cgi_extension .bla
 	}
 
-	location /directory/
+	location /2/ 
 	{
+		root ./tests/sample/
+		index second.html
+		auto_index on
+		auth_key juyang:1234
+	}
+
+	location /virtual/
+	{
+		root ./tests/sample/
+		cgi_extension .bla
+		auth_key kilee:1234
+	}
+
+	location /post/
+	{
+		allow_methods GET POST PUT DELETE
+		root ./tests/sample/tmp/
+		client_body_buffer_size 1000
+		auth_key jayun:1234
+	}
+
+	location /post/tmp/
+	{
+		allow_methods POST PUT
+		root ./test/sample/tmp/
+		upload_path /tmp/
+		auth_key hski:1234
+	}
+
+	location /index/
+	{
+		root ./tests/sample/
 		allow_methods GET
-		root ./tester/YoupiBanane/
-		index youpi.bad_extension
+		return 301 https://www.naver.com
 	}
 }
