--- tests/sample/test.conf_
@@ -1,83 +1,0 @@
-#주석가능하긴한데 예약어로 쓰인 것들은 적으면 안됩니다. 예약어 기준으로 파싱합니다.
-#얌체같이 파싱했습니다.
-#로케이션 중에 루트 로케이션 ( / ) 는 사용하지 않더라도반드시 필요합니다. 제발 써주세요.
-
-server
-{
-	server_name default_server
-	listen 8580 127.0.0.1			# 7번라인 8번라인 반드시 연달아와야합니다.
-
-	location /						# / 로케이션은 반드시 존재해야합니다. (디폴트서버)
-	{
-		error_page 404 ./tests/www/error404.html
-		error_page 405 ./tests/www/error405.html
-		allow_methods GET POST
-		request_max_body_size	2048
-		root ./tests/www/
-		index index.html index2.html
-		auto_index on
-		cgi_info .bla
-		auth_key hski:1234
-	}
-
-	location /2/ 
-	{
-		root ./tests/www/
-		index second.html
-		auto_index on
-		auth_key juyang:1234
-	}
-
-	location /virtual/
-	{
-		root ./tests/www/
-		cgi_info .bla
-		auth_key kilee:1234
-	}
-
-	location /post/
-	{
-		allow_methods GET POST PUT DELETE
-		root ./tests/www/tmp/
-		client_body_buffer_size 1000
-		auth_key jayun:1234
-	}
-
-	location /post/tmp/
-	{
-		allow_methods POST PUT
-		root ./test/www/tmp/
-		upload_path /tmp/
-		auth_key hski:1234
-	}
-
-	location /index/
-	{
-		root ./tests/www/
-		return 301 https://www.naver.com
-	}
-}
-
-server
-{
-	server_name special_server
-	listen 8581 127.0.0.1
-	
-	location /
-	{
-		root ./tests/www/
-		index second.html
-	}
-}
-
-server
-{
-	server_name google.com
-	listen 8582 127.0.0.1
-
-	location /
-	{
-		root ./tests/www/
-		index google.html
-	}
-}
