+++ tests/sample/test__.conf
@@ -5,23 +5,24 @@
 server
 {
 	server_name default_server
-	listen 8180 127.0.0.1			# 7번라인 8번라인 반드시 연달아와야합니다.
+	listen 8580 127.0.0.1			# 7번라인 8번라인 반드시 연달아와야합니다.
 
 	location /						# / 로케이션은 반드시 존재해야합니다. (디폴트서버)
 	{
-		error_page 404 ./tests/www/error404.html
-		error_page 405 ./tests/www/error405.html
+		error_page 404 ./tests/sample/error404.html
+		error_page 405 ./tests/sample/error405.html
 		allow_methods GET POST
 		request_max_body_size	2048
-		root ./tests/www/
+		root ./tests/sample/
 		index index.html index2.html
 		auto_index on
-		cgi_extension .bla
+		cgi_info .bla
+		auth_key hski:1234
 	}
 
 	location /2/ 
 	{
-		root ./tests/www/
+		root ./tests/sample/
 		index second.html
 		auto_index on
 		auth_key juyang:1234
@@ -29,15 +30,15 @@ server
 
 	location /virtual/
 	{
-		root ./tests/www/
-		cgi_extension .bla
+		root ./tests/sample/
+		cgi_info .bla
 		auth_key kilee:1234
 	}
 
 	location /post/
 	{
 		allow_methods GET POST PUT DELETE
-		root ./tests/www/tmp/
+		root ./tests/sample/tmp/
 		client_body_buffer_size 1000
 		auth_key jayun:1234
 	}
@@ -45,14 +46,38 @@ server
 	location /post/tmp/
 	{
 		allow_methods POST PUT
-		root ./test/www/tmp/
+		root ./test/sample/tmp/
 		upload_path /tmp/
 		auth_key hski:1234
 	}
 
 	location /index/
 	{
-		root ./tests/www/
+		root ./tests/sample/
 		return 301 https://www.naver.com
+	}
+}
+
+server
+{
+	server_name special_server
+	listen 8581 127.0.0.1
+	
+	location /
+	{
+		root ./tests/sample/
+		index second.html
+	}
+}
+
+server
+{
+	server_name google.com
+	listen 8582 127.0.0.1
+
+	location /
+	{
+		root ./tests/sample/
+		index google.html
 	}
 }
