+++ tests/test1_tester/test1_tester.config
@@ -10,7 +10,7 @@ server
 	location /					# / 로케이션은 반드시 존재해야합니다. (디폴트서버)
 	{
 		allow_methods GET
-		root ./tests/test1_tester/test1_tester_root
+		root ./tests/test1_tester/test1_tester_root/YoupiBanane/
 		index youpi.bad_extension
 	}
